+++ Quaver.API/Maps/Qua.cs
@@ -264,9 +264,9 @@ namespace Quaver.API.Maps
             switch (Mode)
             {
                 case GameMode.Keys4:
-                    return new StrainSolverKeysNEW(this, new StrainConstantsKeys(), mods);
+                    return new StrainSolverKeys(this, new StrainConstantsKeys(), mods);
                 case GameMode.Keys7:
-                    return new StrainSolverKeysNEW(this, new StrainConstantsKeys(), mods);
+                    return new StrainSolverKeys(this, new StrainConstantsKeys(), mods);
                 default:
                     throw new InvalidEnumArgumentException();
             }
